Add README and base Hooks class
A README file has been added to provide an overview of the project. This includes a brief description, setup instructions, and details on the use of Cucumber and Selenium. In addition, the Hooks class, a basic setup and teardown for tests, has been added under the step definitions. This file configures the WebDriver before each test and quits it after each test.