Rename package from 'texting' to 'testing'
The main change in this commit is the renaming of the package in multiple files from 'texting' to 'testing' to correct a possible typo. This change not only affects the structure of the project, but any references to the package within the code. Also, some minor changes were made in the pom.xml and the addition of a new file "testng.xml".